requirements.txt
awswrangler==3.14.0
boto3==1.41.2
dbt-core==1.10.15
dbt-snowflake==1.10.3
dbt-adapters==1.20.1 
dbt-common==1.36.0 
dbt-extractor==0.6.0 
dbt-protos==1.0.397 
dbt-semantic-interfaces==0.9.0
google-api-python-client==2.187.0
google-api-core==2.28.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
numpy==2.3.5
pandas==2.3.3
psycopg2-binary==2.9.11
pyarrow==21.0.0
snowflake-connector-python==3.18.0       
SQLAlchemy==1.4.49 
Flask-SQLAlchemy==2.5.1
apache-airflow-providers-postgres==5.7.0 
apache-airflow-providers-airbyte
